added router to react & completely restructured directory layout
