+++ frontend/package.json
@@ -14,6 +14,7 @@
     "react-dom": "^16.12.0",
     "react-google-button": "^0.7.0",
     "react-redux": "^7.1.3",
+    "react-router-dom": "^5.1.2",
     "react-scripts": "3.3.0",
     "react-ticker": "^1.2.2",
     "tsutils": "^3.17.1"
