+++ frontend/src/components/articles/NewsFeedRow.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import NewsItem from './NewsItem';
-import articles from '../data/articles';
+import articles from '../../data/articles';
 import GridList from '@material-ui/core/GridList';
 import GridListTile from '@material-ui/core/GridListTile';
 import Typography from '@material-ui/core/Typography';
@@ -35,27 +35,22 @@ class NewsFeed extends Component {
 
         return (
             // NewsFeed = ClassName
-            <div style={{ display: 'fixed', flexWrap: 'nowrap',  overflow: 'show', backgroundColor: 'white' }}>
+            <div style={{ display: 'fixed', flexWrap: 'nowrap', overflow: 'show', backgroundColor: 'white' }}>
                 <Typography>{this.props.newsFeedRow}</Typography>
-                
                 <div>
-
-                    <GridList cols={5} style={{flexWrap: 'nowrap', transform: 'translateZ(0)' }}>
-
+                    <GridList cols={5} style={{ flexWrap: 'nowrap', transform: 'translateZ(0)' }}>
                         {
                             articles.Articles.map((article, i) => {
                                 return (
-                                    <GridListTile style={{height: '220px', width: '250px', margin: '10px'}} key={this.props.i}>
+                                    <GridListTile style={{ height: '220px', width: '250px', margin: '10px' }} key={this.props.i}>
 
                                         <NewsItem key={this.props.i} Article={article} />
-                                        
+
                                     </GridListTile>
                                 )
                             })
                         }
-
                     </GridList>
-
                 </div>
             </div>
