+++ frontend/src/components/pages/Category.js
@@ -1,0 +1,15 @@
+import React, { Component } from 'react';
+
+class Category extends Component {
+
+
+    render() {
+        return (
+            <div>
+                Category Page
+            </div>
+        )
+    }
+}
+
+export default Category
